move some checks to own module
